add blocks

+++ common/triangle.cpp
@@ -1,11 +1,11 @@
 #include<common/triangle.hpp>
-Triangle::Triangle(const glm::mat3 vertices) :vertices(vertices) {}
-glm::mat3 Triangle::GetVertices()const { return vertices; }
-Triangle Triangle::ApplyTransform(const glm::mat4& transform)const {
-	glm::mat3 ret;
-	const glm::mat3& t = GetVertices();
+Triangle::Triangle(const glm::dmat3 vertices) :vertices(vertices) {}
+glm::dmat3 Triangle::GetVertices()const { return vertices; }
+Triangle Triangle::ApplyTransform(const glm::dmat4& transform)const {
+	glm::dmat3 ret;
+	const glm::dmat3& t = GetVertices();
 	for (int i = 0; i < 3; i++) {
-		auto res = (transform * glm::vec4(t[i], 1.0f));
+		auto res = (transform * glm::dvec4(t[i], 1.0f));
 		assert(!glm::any(glm::isnan(res)));
 		ret[i].x = res.x;
 		ret[i].y = res.y;
@@ -13,16 +13,16 @@ Triangle Triangle::ApplyTransform(const glm::mat4& transform)const {
 	}
 	return Triangle(ret);
 }
-std::vector<Triangle>Triangle::Cube(const glm::vec3 &xyz) {
-	glm::vec3 ps[8] = {
-		glm::vec3(-xyz.x,-xyz.y, xyz.z),
-		glm::vec3( xyz.x,-xyz.y, xyz.z),
-		glm::vec3(-xyz.x,-xyz.y,-xyz.z),
-		glm::vec3( xyz.x,-xyz.y,-xyz.z),
-		glm::vec3(-xyz.x, xyz.y, xyz.z),
-		glm::vec3( xyz.x, xyz.y, xyz.z),
-		glm::vec3(-xyz.x, xyz.y,-xyz.z),
-		glm::vec3( xyz.x, xyz.y,-xyz.z),
+std::vector<Triangle>Triangle::Cube(const glm::dvec3 &xyz) {
+	glm::dvec3 ps[8] = {
+		glm::dvec3(-xyz.x,-xyz.y, xyz.z),
+		glm::dvec3( xyz.x,-xyz.y, xyz.z),
+		glm::dvec3(-xyz.x,-xyz.y,-xyz.z),
+		glm::dvec3( xyz.x,-xyz.y,-xyz.z),
+		glm::dvec3(-xyz.x, xyz.y, xyz.z),
+		glm::dvec3( xyz.x, xyz.y, xyz.z),
+		glm::dvec3(-xyz.x, xyz.y,-xyz.z),
+		glm::dvec3( xyz.x, xyz.y,-xyz.z),
 	};
 	/*
 	  o----o
@@ -33,17 +33,17 @@ std::vector<Triangle>Triangle::Cube(const glm::vec3 &xyz) {
 	o----o
 	*/
 	return {
-		glm::mat3(ps[0],ps[2],ps[3]),
-		glm::mat3(ps[0],ps[3],ps[1]),
-		glm::mat3(ps[0],ps[1],ps[5]),
-		glm::mat3(ps[0],ps[5],ps[4]),
-		glm::mat3(ps[0],ps[4],ps[6]),
-		glm::mat3(ps[0],ps[6],ps[2]),
-		glm::mat3(ps[7],ps[6],ps[4]),
-		glm::mat3(ps[7],ps[4],ps[5]),
-		glm::mat3(ps[7],ps[5],ps[1]),
-		glm::mat3(ps[7],ps[1],ps[3]),
-		glm::mat3(ps[7],ps[3],ps[2]),
-		glm::mat3(ps[7],ps[2],ps[6])
+		glm::dmat3(ps[0],ps[2],ps[3]),
+		glm::dmat3(ps[0],ps[3],ps[1]),
+		glm::dmat3(ps[0],ps[1],ps[5]),
+		glm::dmat3(ps[0],ps[5],ps[4]),
+		glm::dmat3(ps[0],ps[4],ps[6]),
+		glm::dmat3(ps[0],ps[6],ps[2]),
+		glm::dmat3(ps[7],ps[6],ps[4]),
+		glm::dmat3(ps[7],ps[4],ps[5]),
+		glm::dmat3(ps[7],ps[5],ps[1]),
+		glm::dmat3(ps[7],ps[1],ps[3]),
+		glm::dmat3(ps[7],ps[3],ps[2]),
+		glm::dmat3(ps[7],ps[2],ps[6])
 	};
 }
